pyproject.toml
[tool.poetry]
name = "marketplace"
version = "0.1.0"
description = "Marketplace engine"
authors = ["Weni"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.4"
django-environ = "0.4.5"
psycopg2 = "^2.8.6"
boto3 = "^1.17.107"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
mozilla-django-oidc = "^2.0.0"
drf-nested-routers = "^0.93.3"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
flake8 = "^3.9.2"
coverage = "^5.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
